version 1.2
Submitted it to marketplace but had a number of crashes. The bug came
when deleting ideas. Fixed that bug and added functionality of pressing
and holding an idea on the main page to show a menu offering to pin idea
to start, share it or delete it. 100 downloads in 10 days. That's good,
right?